+++ PBQLOutlineView.m
@@ -62,11 +62,11 @@
 - (BOOL) outlineView: (NSOutlineView *)ov
          isItemExpandable: (id)item { return NO; }
 
-- (int)  outlineView: (NSOutlineView *)ov
+- (NSInteger)  outlineView: (NSOutlineView *)ov
          numberOfChildrenOfItem:(id)item { return 0; }
 
 - (id)   outlineView: (NSOutlineView *)ov
-         child:(int)index
+         child:(NSInteger)index
          ofItem:(id)item { return nil; }
 
 - (id)   outlineView: (NSOutlineView *)ov
